requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
Flask-Cors==3.0.8
furl==2.1.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.2
orderedmultidict==1.0.1
pika==1.1.0
Pillow==7.0.0
protobuf==3.11.3
pycodestyle==2.5.0
pyflakes==2.1.1
pymongo==3.10.1
pynpm==0.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
virtualenv==16.7.7
Werkzeug==1.0.1
